# docs/swagger.yaml
basePath: /
definitions:
...
  model.AuthServiceUnavailableErrorDetail:
    properties:
      code:
        example: auth_service_unavailable
        type: string
    type: object
...
  model.CreateUserRequest:
    properties:
      hak_akses_ids:
        items:
          type: integer
        type: array
      id_departemen:
        type: integer
      id_mitra:
        type: integer
      is_manager:
        type: boolean
      password:
        minLength: 6
        type: string
      status:
        description: Opsional, default 'active'
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  model.GetMeUnauthorizedDoc:
    properties:
      error: {}
      message:
        example: unauthorized
        type: string
      status:
        example: error
        type: string
    type: object
...
  model.LoginBadRequestDoc:
    properties:
      error: {}
      message:
        example: bad request
        type: string
      status:
        example: error
        type: string
    type: object
...
  model.LoginServiceUnavailableDoc:
    properties:
      error:
        $ref: '#/definitions/model.AuthServiceUnavailableErrorDetail'
      message:
        example: authentication service unavailable
        type: string
      status:
        example: error
        type: string
    type: object
...
  model.UpdateUserRequest:
    properties:
      hak_akses_ids:
        items:
          type: integer
        type: array
      id_departemen:
        type: integer
      id_mitra:
        type: integer
      is_manager:
        type: boolean
      password:
        minLength: 6
        type: string
      status:
        type: string
      username:
        type: string
    required:
    - username
    type: object
  model.UserListSuccessDoc:
    properties:
      data:
        items:
          $ref: '#/definitions/model.UserResponse'
        type: array
      message:
        example: users retrieved
        type: string
      status:
        example: success
        type: string
    type: object
  model.UserResponse:
    properties:
      created_at:
        type: string
      hak_akses_ids:
        items:
          type: integer
        type: array
      id_departemen:
        type: integer
      id_mitra:
        type: integer
      id_user:
        type: integer
      is_manager:
        type: boolean
      nama_departemen:
        type: string
      nama_perusahaan:
        type: string
      permissions:
        items:
          type: string
        type: array
      status:
        type: string
      username:
        type: string
    type: object
  model.UserSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.UserResponse'
      message:
        example: user created
        type: string
      status:
        example: success
        type: string
    type: object
...
  response.BaseResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@permatatex.local
    name: API Support
  description: API documentation for Permatatex IT Inventory System backend.
  license:
    name: MIT
  termsOfService: https://swagger.io/terms/
  title: Permatatex IT Inventory API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      description: Returns a paginated list of users with their department and mitra
        names.
      parameters:
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserListSuccessDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.GetMeUnauthorizedDoc'
      security:
      - BearerAuth: []
      summary: List Users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Creates a new user and assigns permissions in a single transaction.
      parameters:
      - description: Create user payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.UserSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.LoginBadRequestDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.GetMeUnauthorizedDoc'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.LoginBadRequestDoc'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/model.LoginServiceUnavailableDoc'
      security:
      - BearerAuth: []
      summary: Create New User
      tags:
      - Users
  /api/v1/users/{id}:
    delete:
      description: Deletes a user by ID. Protects Super Admin (ID 1) from deletion.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.LoginBadRequestDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.GetMeUnauthorizedDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.LoginBadRequestDoc'
      security:
      - BearerAuth: []
      summary: Delete User
      tags:
      - Users
    get:
      description: Returns a single user detail with permissions.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserSuccessDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.GetMeUnauthorizedDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.LoginBadRequestDoc'
      security:
      - BearerAuth: []
      summary: Get User Detail
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Updates user profile and permissions in a single transaction. Password
        is optional.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update user payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.LoginBadRequestDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.GetMeUnauthorizedDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.LoginBadRequestDoc'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.LoginBadRequestDoc'
      security:
      - BearerAuth: []
      summary: Update User
      tags:
      - Users
  /api/v1/users/{id}/approve:
    put:
      description: Approves a pending user registration, updating status to active.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserSuccessDoc'
      security:
      - BearerAuth: []
      summary: Approve User Pendaftaran
      tags:
      - Users
  /api/v1/users/{id}/reject:
    put:
      description: Rejects a pending user registration, updating status to rejected.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - BearerAuth: []
      summary: Reject User Pendaftaran
      tags:
      - Users
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"